Enhance G-Purchasing

+++ src/modules/garment-purchasing/customs/view.js
@@ -84,17 +84,27 @@ export class View {
   }
 
   cancel(event) {
-    this.router.navigateToRoute('list');
+    var r = confirm("Apakah anda yakin akan Keluar?")
+    if ( r == true) {
+      this.router.navigateToRoute('list');
+    }
+    
   }
 
   edit(event) {
-    this.router.navigateToRoute('edit', { id: this.data._id });
+    var r = confirm("Apakah anda yakin akan mengubah data ini?")
+    if (r == true) {
+      this.router.navigateToRoute('edit', { id: this.data._id });
+    }
   }
 
   delete(event) {
-    this.service.delete(this.data)
+    var r = confirm("Apakah anda yakin akan menghapus data ini?")
+    if (r == true) {
+      this.service.delete(this.data)
       .then(result => {
         this.cancel();
       });
+    }
   }
 }
